Finished testing out code for retrieving and storing an entire season. Changed the _ID_exists_in_DB function a bit for now, but will eventually interact with the sql database.

+++ scrape_core.py
@@ -9,9 +9,10 @@ import ipdb
 
 class Scrape(object):
 
-    def __init__(self,year=None):
+    def __init__(self,year=None,debug=False):
         self.year = year
         self.base_url = 'http://www.basketball-reference.com'
+        self.debug=debug
         
     def _check_game_exists(self, boxscore_link):
         """
@@ -32,8 +33,12 @@ class Scrape(object):
         downloaded JSON files, assuming they are stored in ../../data/ .
         """
         PATH='../../data/'
-
-        return os.path.exists(PATH+ID+'.json')
+        files = os.listdir(PATH)
+        exists = False
+        for f in files:
+            if ID in f:
+                exists = True
+        return exists
     
     def _get_game_urls_for_season(self):
         """
@@ -72,7 +77,7 @@ class Scrape(object):
         game_url : str
             The basketball-reference.com game url.
         field_type : str
-            'basic' or 'advanced', pertaining to the type of data obtained
+            'basic' or 'advanced', pertaining to the field_type of data obtained
 
         Returns
         -------
@@ -97,34 +102,39 @@ class Scrape(object):
                                    'AST%','STL%','BLK%','TOV%','USG%','ORtg',
                                    'DRtg']}
 
-        # Obtain the html
-        html = urlopen(self.game_url).read()
+        # Check for game data existence in db
         gametag = self.game_url.split('/')[-1].strip('.html')
-        soup = BeautifulSoup(html,'html.parser')
+        exists = self._ID_exists_in_DB(gametag+field_type)
+        # Obtain the html if game doesn't exist
+        if exists is False:
+            html = urlopen(self.game_url).read()
+            soup = BeautifulSoup(html,'html.parser')
+            fields = self.data_fields['{}'.format(field_type)]
         
-        fields = self.data_fields['{}'.format(field_type)]
-        
-        # Get table from html
-        tables = soup.findAll('table', id=re.compile(field_type))
-        tables = [t.find('tbody') for t in tables]
+            # Get table from html
+            tables = soup.findAll('table', id=re.compile(field_type))
+            tables = [t.find('tbody') for t in tables]
     
-        for t in tables:
-            data = t.findAll('tr')
-            for player in data:
-                if player['class'][0] == '':
-                    pdata = player.findAll('td')
-                    name = pdata[0].string
-                    # is this sufficient to always determine if
-                    # the player wasn't on the bench?
-                    if len(pdata) > 2: 
-                        tmp = self.boxscore_data.get(name, {})
-                        for i,k in enumerate(fields):
-                            if pdata[i] == None:
-                                pdata[i] = 'None'
-                            tmp[k] = pdata[i+1].string
-                        self.boxscore_data[name] = tmp
+            for t in tables:
+                data = t.findAll('tr')
+                for player in data:
+                    if player['class'][0] == '':
+                        pdata = player.findAll('td')
+                        name = pdata[0].string
+                        # is this sufficient to always determine if
+                        # the player wasn't on the bench?
+                        if len(pdata) > 2: 
+                            tmp = self.boxscore_data.get(name, {})
+                            for i,k in enumerate(fields):
+                                if pdata[i] == None:
+                                    pdata[i] = 'None'
+                                tmp[k] = pdata[i+1].string
+                            self.boxscore_data[name] = tmp
 
-        return self.boxscore_data, gametag
+            return self.boxscore_data, gametag
+        else:
+            print("Game {} already exists in db. Skipping web-scraping phase...".format(gametag))
+            return self.boxscore_data, gametag
 
     def display_boxscore(self,boxscore_data):
         """
@@ -133,7 +143,7 @@ class Scrape(object):
         Parameters
         ----------
         boxscore_data : dict
-            A dictionary of data for every player. Basic/Advanced type is learned
+            A dictionary of data for every player. Basic/Advanced field_type is learned
         from the input.
 
         Returns
@@ -141,22 +151,22 @@ class Scrape(object):
         """
         
         data = []
-        type = None
+        field_type = None
         
         # headers
         if len(boxscore_data[[i for i in boxscore_data.keys()][0]]) == 20:
             data.append(["Name"] + self.data_fields['basic'])
-            type='basic'
+            field_type='basic'
         elif len(boxscore_data[[i for i in boxscore_data.keys()][0]]) == 13:
             data.append(["Name"] + self.data_fields['advanced'])
-            type='advanced'
+            field_type='advanced'
         else:
             "ERROR MESSAGE HERE"
 
         # data
         for player in self.boxscore_data:
             row = [player]
-            for k in self.data_fields['{}'.format(type)]:
+            for k in self.data_fields['{}'.format(field_type)]:
                 row.append(self.boxscore_data[player][k])
             data.append(row)
         print(tabulate(data))
@@ -171,33 +181,60 @@ class Scrape(object):
         Parameters
         ----------
         boxscore_data : dict
-            A dictionary of data for every player. Basic/Advanced type is learned
+            A dictionary of data for every player. Basic/Advanced field_type is learned
             from the input.
         gametag : str
             Unique game id number.
         """
-
-        # Figures out if type is 'basic' or 'advanced'
-        type = None
+        
+        # Checks for the empty dictionary, which is an indication that the data have
+        # not been retrieved from the web; it already exists locally.
+        field_type = None
+        if len(boxscore_data.keys()) == 0:
+            if self.debug is True:
+                print("Game {} already exists in db. Skipping storage phase...".format(gametag))
+            return
+        # Figures out if field_type is 'basic' or 'advanced'        
         if len(boxscore_data[[i for i in boxscore_data.keys()][0]]) == 20:
-            type='basic'
+            field_type='basic'
         elif len(boxscore_data[[i for i in boxscore_data.keys()][0]]) == 13:
-            type='advanced'
+            field_type='advanced'
         else:
-            "ERROR MESSAGE HERE"
+            print("ERROR: Neither basic nor advanced data is being requested...")
 
-        with open ('../../data/'+gametag+'.json','w') as f:
-            json.dump(boxscore_data, f)
+        # Check to see if game_data exists in database
+        exists = self._ID_exists_in_DB(gametag+field_type)
+        if exists is False:
+            with open ('../../data/'+gametag+field_type+'.json','w') as f:
+                json.dump(boxscore_data, f)
+            if self.debug is True:
+                print("Wrote game {} out to file...".format(gametag))
+        else:
+            if self.debug is True:
+                print("Game {} already exists in db. Skipping storage phase...".format(gametag))
         
 if __name__ == "__main__":
     
     # Test #1: Scrape 2014, and choose the first game.
     # Get boxscore, and print table to terminal. Then
     # dump game data to json file (currently on Desktop).
+    '''
     s2014 = Scrape(year=2014)
     links2014 = s2014._get_game_urls_for_season()
     boxscore_data,gametag = s2014._get_boxscore(links2014[0],field_type='basic')
     s2014.display_boxscore(boxscore_data)
     s2014._dump_gamedata_to_json(boxscore_data,gametag)
-    
+    '''
+
+    # Test #2: Store game data for all basketball games
+    # for 2013 season. Check to see if any have been
+    # dumped to json files previously. Eventually, check
+    # against database rather than json file.
+    s2013 = Scrape(year=2013,debug=True)
+    links2013 = s2013._get_game_urls_for_season()
+    for li in links2013:
+        boxscore_data,gametag = s2013._get_boxscore(li,field_type='advanced')
+        s2013._dump_gamedata_to_json(boxscore_data,gametag)
+        boxscore_data,gametag = s2013._get_boxscore(li,field_type='basic')
+        s2013._dump_gamedata_to_json(boxscore_data,gametag)
